requirements.txt
pandas
psycopg2-binary
sshtunnel
python-dotenv
tqdm
sphinx
furo
setuptools